FileSystemAdapter.copyStreamToFile() supports dry run option

+++ src/main/java/jdrivesync/fs/FileSystemAdapter.java
@@ -180,12 +180,14 @@ public class FileSystemAdapter {
 	}
 
 	private void copyStreamToFile(InputStream inputStream, java.io.File file) throws IOException {
-		byte[] buffer = new byte[1024];
-		try (FileOutputStream fos = new FileOutputStream(file)) {
-			int read = inputStream.read(buffer, 0, buffer.length);
-			while (read >= 0) {
-				fos.write(buffer, 0, read);
-				read = inputStream.read(buffer, 0, buffer.length);
+		if (options.isDryRun()) {
+			byte[] buffer = new byte[1024];
+			try (FileOutputStream fos = new FileOutputStream(file)) {
+				int read = inputStream.read(buffer, 0, buffer.length);
+				while (read >= 0) {
+					fos.write(buffer, 0, read);
+					read = inputStream.read(buffer, 0, buffer.length);
+				}
 			}
 		}
 	}
